refactor: Replace useAutoLayout with useInitialAutoLayout for improved layout handling on node initialization
